ci: start test containter from code
