--- docker-compose.yaml
@@ -1,13 +1,0 @@
-version: "3.4"
-services:
-  rabbitmq:
-      container_name: rabbitmq
-      image: 'rabbitmq:3.13.3-management'
-      ports:
-        - "5672:5672"
-        - "15672:15672"
-      healthcheck:
-        test: rabbitmq-diagnostics -q ping
-        interval: 5s
-        timeout: 10s
-        retries: 5
